ebtables: replace ebtables-save perl script with bash rewrite
Fedora provides a bash replacement for the default ebtables-save perl
script. Using it allows the ebtables run-time dependency on perl to
be replaced with a runtime dependency on bash - which is lower
overhead and more likely to be present on typical embedded systems
already.

  https://bugzilla.redhat.com/show_bug.cgi?id=746040
  http://pkgs.fedoraproject.org/cgit/rpms/ebtables.git/tree/ebtables-save

Since ebtables-save no longer contains a reference to /usr, the
previous QA issue workaround of moving it from ${base_sbindir} to
${sbindir} is no longer required.

  http://git.openembedded.org/meta-openembedded/commit/?id=a7c6fcebee7d9f86c356ea92de445d89e714ff62

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>