hostapd: fix LICENSE + misc recipe updates
According to the COPYING file in the top level of the hostapd source
tree, hostapd was relicensed from dual BSD/GPLv2 to BSD only in
February 2012. This change has apparently gone un-noticed for the
past 6 years, but fix it now. Also make some minor formatting tweaks
to bring the recipe more in line with the OE Styleguide.

Also use pkg-config to find libnl headers (instead of hardcoding),
append to base do_configure (instead of over-riding) and respect OE's
default CFLAGS.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>