nano: update from 8.3 -> 8.4
Release notes:
   https://lists.gnu.org/archive/html/info-gnu/2025-04/msg00001.html

The license file checksum changed because upstream added the line:
   SPDX-License-Identifier: BSD-3-Clause
Add OECONF to avoid build-time checks for strcase[str|cmp], both of
which have been in glibc and musl for > 10 years so it's sensible to
default to yes.

Adjust the patches using devtool and manually munge the Upstream-Status-es.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>